pyproject.toml
[project]
name = "pdf-suite"
version = "0.1.0"
description = "Full-featured PDF editing suite for Frappe"
authors = [
    {name = "Codebase Technologies", email = "dev@codebtech.com"}
]
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pikepdf>=8.0.0",
    "pdfplumber>=0.10.0",
    "reportlab>=4.0",
    "fpdf2>=2.7.0",
    "pytesseract>=0.3.10",
    "pdf2docx>=0.5.8",
    "python-docx>=1.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0"

[tool.bench.apt-packages]
packages = [
    "tesseract-ocr",
    "tesseract-ocr-eng",
    "tesseract-ocr-ara",
    "libreoffice-writer",
    "poppler-utils",
    "ghostscript",
]
